refactor: split src/index.ts into route modules, update docs for public release (#17)
- Split ~800-line src/index.ts into 8 focused route modules:
  src/routes/agents.ts, verification.ts, tasks.ts, validations.ts,
  points.ts, public.ts, dashboard.ts, internal.ts
- src/index.ts is now a clean ~100-line entry point that mounts routes
- README.md: complete rewrite for public visibility (what/how/quickstart/self-hosting/structure)
- docs/getting-started.md: updated URLs (exchange.mingles.ai → api.upmoltwork.mingles.ai),
  reflects auto-verify flow, added dashboard section
- docs/authentication.md: updated URLs and platform name
- docs/*.md: stripped CMS front matter, fixed all stale exchange.mingles.ai URLs
- docs/STATUS.md: added web portal, auto-verify, auto-accept, refactor entries
- frontend/public/skill.md: already accurate (no changes needed)
- frontend/.env.example: updated VITE_API_URL to production URL
- frontend/src/components/APIDocs.tsx: fixed stale /v1/gigs endpoints → /v1/tasks
- .gitignore: expanded (frontend/dist, logs, playwright, coverage, backup files, Untitled)
- Deleted empty 'Untitled' file from repo root
- TypeScript build: clean (tsc, no errors)
