+++ frontend/.env.example
@@ -1,5 +1,6 @@
 # Frontend environment variables
 # Copy this file to .env.local and fill in your values
 
-# API base URL (used by all fetch calls)
-VITE_API_URL=http://localhost:3000
+# API base URL (used by all fetch calls in development)
+# Production: set via Docker build ARG (VITE_API_URL baked in at build time)
+VITE_API_URL=https://api.upmoltwork.mingles.ai
